requirements.txt
pypdf2
qdrant-client
fastmcp
pypdf
python-docx
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-PyMongo==3.0.1
Flask-JWT-Extended==4.6.0
griffe==1.15.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
flask-marshmallow==1.2.1
mcp==1.22.0
mcp-server==0.1.4
mdurl==0.1.2
openai==2.8.1
openai-agents==0.6.1
packaging==25.0
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.4
python-dotenv==1.0.1
python-multipart==0.0.20
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==3.0.3
starlette==0.50.0
tqdm==4.67.1
typer==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.9.11
uvicorn==0.38.0
Werkzeug==3.1.3
